Refactor: Align reverse-linked-list structure with 3sum
Moves the 'reverse-linked-list' problem from a single file (`packages/backend/src/problem/free/reverse-linked-list.ts`) into a directory structure (`packages/backend/src/problem/free/reverse-linked-list/`) consistent with other problems like '3sum'.

This involves:
- Creating separate files for types, code, steps, problem definition, and test cases (`types.ts`, `code/typescript.ts`, `steps.ts`, `problem.ts`, `testcase.ts`).
- Adding standard placeholder files (`index.test.ts`, `variables.ts`, `groups.ts`).
- Updating all imports and exports to reflect the new structure.
- Creating an `index.ts` to export the main problem definition.
- Deleting the original single file.

This refactoring improves consistency and organization within the problems directory.
