+++ packages/backend/src/problem/free/reverse-linked-list/index.test.ts
@@ -1,0 +1,1 @@
+// Empty file - tests will be added later if needed.
