+++ packages/backend/src/problem/free/reverse-linked-list/types.ts
@@ -1,0 +1,5 @@
+import { ListNode } from "algo-lens-core";
+
+export interface ReverseListInput {
+  head: ListNode | null;
+}
