+++ packages/backend/src/problem/free/reverse-linked-list/variables.ts
@@ -1,0 +1,1 @@
+// Empty file - will be populated later if needed.
